requirements.txt
aiohttp==3.7.4.post0
aniso8601==9.0.1
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.4.0
autopep8==2.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
configparser==5.3.0
contourpy==1.0.6
cssselect==1.1.0
cycler==0.11.0
dataframe-image==0.1.3
DateTime==4.7
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.4
fake-useragent==1.1.1
fastapi==0.78.0
fastjsonschema==2.16.2
fonttools==4.38.0
h11==0.14.0
html2image==2.0.1
httpcore==0.16.2
httpx==0.23.1
idna==3.3
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
mistune==2.0.4
multidict==6.0.2
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parse==1.19.0
Pillow==9.3.0
platformdirs==2.6.2
pycodestyle==2.10.0
pydantic==1.9.1
pyee==8.2.2
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.28.1
requests-html==0.10.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.19.1
tabulate==0.9.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
w3lib==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
yarl==1.8.2
zipp==3.11.0
zope.interface==5.5.2
